Move unittests to a separate file
