+++ yt/frontends/ytdata/tests/test_old_outputs.py
@@ -27,17 +27,12 @@ from yt.frontends.ytdata.tests.test_outputs import \
 from yt.testing import \
     assert_allclose_units, \
     assert_array_equal, \
-    assert_fname, \
-    requires_file, \
-    requires_module
+    requires_file
 from yt.utilities.answer_testing.framework import \
     requires_ds, \
     data_dir_load
 from yt.units.yt_array import \
     YTArray
-from yt.visualization.plot_window import \
-    SlicePlot, \
-    ProjectionPlot
 from yt.visualization.profile_plotter import \
     ProfilePlot, \
     PhasePlot
@@ -184,36 +179,3 @@ def test_old_nonspatial_data():
     new_ds = data_dir_load(full_fn)
     assert isinstance(new_ds, YTNonspatialDataset)
     yield YTDataFieldTest(full_fn, "density", geometric=False)
-
-@requires_module('h5py')
-@requires_file(os.path.join(ytdata_dir, "slice.h5"))
-@requires_file(os.path.join(ytdata_dir, "proj.h5"))
-@requires_file(os.path.join(ytdata_dir, "oas.h5"))
-def test_old_plot_data():
-    tmpdir = tempfile.mkdtemp()
-    curdir = os.getcwd()
-    os.chdir(tmpdir)
-
-    fn = "slice.h5"
-    full_fn = os.path.join(ytdata_dir, fn)
-    ds_slice = data_dir_load(full_fn)
-    p = SlicePlot(ds_slice, 'z', 'density')
-    fn = p.save()
-    assert_fname(fn[0])
-
-    fn = "proj.h5"
-    full_fn = os.path.join(ytdata_dir, fn)
-    ds_proj = data_dir_load(full_fn)
-    p = ProjectionPlot(ds_proj, 'z', 'density')
-    fn = p.save()
-    assert_fname(fn[0])
-
-    fn = "oas.h5"
-    full_fn = os.path.join(ytdata_dir, fn)
-    ds_oas = data_dir_load(full_fn)
-    p = SlicePlot(ds_oas, [1, 1, 1], 'density')
-    fn = p.save()
-    assert_fname(fn[0])
-
-    os.chdir(curdir)
-    shutil.rmtree(tmpdir)
