+++ yt/frontends/ytdata/tests/test_outputs.py
@@ -12,10 +12,7 @@ from yt.frontends.ytdata.api import \
 from yt.testing import \
     assert_array_equal, \
     assert_allclose_units, \
-    assert_equal, \
-    assert_fname, \
-    fake_random_ds, \
-    requires_module
+    assert_equal
 from yt.utilities.answer_testing.framework import \
     requires_ds, \
     data_dir_load, \
@@ -23,9 +20,6 @@ from yt.utilities.answer_testing.framework import \
 from yt.units.yt_array import \
     YTArray, \
     YTQuantity
-from yt.visualization.plot_window import \
-    SlicePlot, \
-    ProjectionPlot
 from yt.visualization.profile_plotter import \
     ProfilePlot, \
     PhasePlot
@@ -243,38 +237,6 @@ def test_nonspatial_data():
     new_ds = load(full_fn)
     assert isinstance(new_ds, YTNonspatialDataset)
     yield YTDataFieldTest(full_fn, "density", geometric=False)
-    os.chdir(curdir)
-    if tmpdir != '.':
-        shutil.rmtree(tmpdir)
-
-@requires_module('h5py')
-def test_plot_data():
-    tmpdir = make_tempdir()
-    curdir = os.getcwd()
-    os.chdir(tmpdir)
-    ds = fake_random_ds(16)
-
-    plot = SlicePlot(ds, 'z', 'density')
-    fn = plot.data_source.save_as_dataset('slice.h5')
-    ds_slice = load(fn)
-    p = SlicePlot(ds_slice, 'z', 'density')
-    fn = p.save()
-    assert_fname(fn[0])
-
-    plot = ProjectionPlot(ds, 'z', 'density')
-    fn = plot.data_source.save_as_dataset('proj.h5')
-    ds_proj = load(fn)
-    p = ProjectionPlot(ds_proj, 'z', 'density')
-    fn = p.save()
-    assert_fname(fn[0])
-
-    plot = SlicePlot(ds, [1, 1, 1], 'density')
-    fn = plot.data_source.save_as_dataset('oas.h5')
-    ds_oas = load(fn)
-    p = SlicePlot(ds_oas, [1, 1, 1], 'density')
-    fn = p.save()
-    assert_fname(fn[0])
-
     os.chdir(curdir)
     if tmpdir != '.':
         shutil.rmtree(tmpdir)
